requirements.txt
mcp>=0.9.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
langgraph>=0.0.40
streamlit>=1.28.0
prometheus-client>=0.20.0
